requirements.txt
# Backend и API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning
torch==2.1.1
torchvision==0.16.1
transformers==4.36.0
timm==0.9.12
ultralytics==8.0.220
xgboost==2.0.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1
scikit-image==0.22.0
scipy==1.11.4

# Frontend
streamlit==1.28.2
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
requests==2.31.0

# Development
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Jupyter
jupyter==1.0.0
notebook==7.0.6

# Monitoring
wandb==0.16.0 